multiply feature added

+++ src/main/java/com/qa/calculator/SubNumbers.java
@@ -2,7 +2,7 @@ package com.qa.calculator;
 
 public class SubNumbers {
 	
-	public int sub(int num1, int num2) {
+	public double sub(double num1, double num2) {
 		return num1 - num2;
 	}
